refactor: Move bookmark funs to separate file
